requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
requests==2.31.0
python-multipart==0.0.9
python-dotenv==1.0.0
chromadb==0.5.5
jsonschema==4.19.0
rich==13.4.0